requirements.txt
argh==0.26.1
docopt==0.6.1
ecdsa==0.11
Fabric==1.9.0
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
ordereddict==1.1
paramiko==1.15.2
pathtools==0.1.2
pycrypto==2.6.1
Pygments==1.6
PyYAML==3.11
simiki==1.2.4
watchdog==0.8.2
